+++ src/lib/scripts/index.ts
@@ -1,5 +1,9 @@
 import fundWallets from './fundWallets'
+import liquid from './liquid'
+
+export {LiquidOptions} from "./liquid"
 
 export default {
-    fundWallets
+    fundWallets,
+    liquid,
 }
